requirements.txt
fastapi
uvicorn[standard]
python-dotenv
huggingface-hub
langchain-community
faiss-cpu
